WORKS! Solved sequentially with variable array sizes and precision specified by command line arguments.

+++ src/relaxation_context.rs
@@ -1,7 +1,7 @@
 
 
 pub struct RelaxationContext {
-    pub array_size: i64,
+    pub array_size: usize,
     pub target_precision: f64,
-    pub correct_array: Vec<Vec<i64>>
+    pub correct_array: Vec<Vec<f64>>
 }
